stop racing condition where mappers try to delete from DB smthng deleted by another mapper

+++ im-crawler/src/main/java/com/wipro/ats/bdre/imcrawler/model/DocidsDBDao.java
@@ -93,7 +93,9 @@ public class DocidsDBDao {
         session.beginTransaction();
         Long longid = new Long (id.intValue());
         Docidsdb docidsdb=(Docidsdb)session.get(Docidsdb.class,longid);
-        session.delete(docidsdb);
+        //delete only if it is present from before i.e. not deleted by other mapper
+        if(docidsdb != null)
+            session.delete(docidsdb);
         session.getTransaction().commit();
     }
 }
