+++ im-crawler/src/main/java/com/wipro/ats/bdre/imcrawler/model/WebUrlsDBDao.java
@@ -69,7 +69,9 @@ public class WebUrlsDBDao {
         session.beginTransaction();
         Long lid = new Long(id.intValue());
         Weburlsdb weburlsdb=(Weburlsdb)session.get(Weburlsdb.class,lid);
-        session.delete(weburlsdb);
+        //delete only if it is present from before i.e. not deleted by other mapper
+        if(weburlsdb != null)
+            session.delete(weburlsdb);
         session.getTransaction().commit();
     }
 }
